+++ lib/parse_error.dart
@@ -1,0 +1,1 @@
+part 'src/parse_error/parse_error.dart';