+++ lib/parser.dart
@@ -3,10 +3,11 @@ import 'dart:typed_data';
 import 'package:meta/meta.dart';
 
 import 'buffer.dart';
-import 'parse.dart';
+import 'parse_error.dart';
 import 'range.dart';
 import 'tuple.dart';
 
+export 'parse_error.dart';
 export 'range.dart';
 export 'tuple.dart';
 
@@ -19,13 +20,13 @@ part 'src/parser/char_parser.dart';
 part 'src/parser/chars_parser.dart';
 part 'src/parser/choice_parser.dart';
 part 'src/parser/digit_parser.dart';
+part 'src/parser/fail_parser.dart';
 part 'src/parser/left_parser.dart';
 part 'src/parser/many_parser.dart';
 part 'src/parser/not_parser.dart';
 part 'src/parser/opt_parser.dart';
 part 'src/parser/parse_state.dart';
 part 'src/parser/parser.dart';
-part 'src/parser/punct_parser.dart';
 part 'src/parser/range_parser.dart';
 part 'src/parser/rep_parser.dart';
 part 'src/parser/right_parser.dart';
@@ -34,3 +35,4 @@ part 'src/parser/seq_parser.dart';
 part 'src/parser/skip_many_parser.dart';
 part 'src/parser/str_parser.dart';
 part 'src/parser/tuple_parser.dart';
+part 'src/parser/val_parser.dart';